+++ imports/ui/signers/SignersChartContainer.js
@@ -12,7 +12,7 @@ export default SignersChartContainer = withTracker((props) => {
 
     if (Meteor.isClient){
         heightHandle = Meteor.subscribe('blocks.height', props.limit);
-        validatorHandle = Meteor.subscribe('validators.voting_power');
+        validatorHandle = Meteor.subscribe('validators.all');
         loading = !heightHandle.ready() && !validatorHandle.ready();
     }
 
@@ -25,10 +25,12 @@ export default SignersChartContainer = withTracker((props) => {
         blocks = Blockscon.find({}, {sort: {height:-1}}).fetch();
         validators =  Validators.find({
                 status: 2,
-                jailed:false
+                jailed: false
             },{
                 sort:{
-                    voting_power:-1
+                    voting_power:-1,
+                    uptime: -1,
+                    limit: 96
                 }
             }
         ).fetch();
@@ -49,6 +51,6 @@ export default SignersChartContainer = withTracker((props) => {
         blocksExist,
         validatorsExist,
         blocks: blocksExist ? blocks : {},
-        validators: validators
+        validators: validatorsExist ? validators: {}
     };
 })(SignersChart);